+++ src/validators.js
@@ -58,8 +58,10 @@ export const emailOrAccountId = value => {
 export const telegramUsername = value => {
   return /^[a-zA-Z0-9_-]{3,16}$/.test(value)
 }
-export const emailOrPhoneNumber = value => {
-  return validateEmail(value) || validatePhoneNumber(value)
+export const emailOrPhoneNumberOrTelegram = value => {
+  return validateEmail(value) ||
+    validatePhoneNumber(value) ||
+    telegramUsername(value)
 }
 export const accountId = value => {
   return base.Keypair.isValidPublicKey(value)